==> 15-Recommender Systems/02-Recomendacao_Livros.ipynb <==
# Sistema de Recomendação de Livros

Importando bibliotecas

import pandas as pd
import numpy as np
import matplotlib.pyplot as plt
import seaborn as sns

Leitura do arquivo de livros. Sep é o separador do arquivo, on_bad_lines ignora as linhas que não estão no formato correto, encoding é o tipo de codificação do arquivo e low_memory é para que o pandas não precise alocar muita memória para ler o arquivo.

books = pd.read_csv('BX-Books.csv', sep=';',on_bad_lines='skip',encoding="latin-1", low_memory=False)
books.head()

books.shape

É um dataset com 271.360 livros e 8 atributos.

books.columns

- ISBN: Número de identificação do livro
- Book-Title: Título do livro
- Book-Author: Autor do livro
- Year-Of-Publication: Ano de publicação do livro
- Publisher: Editora do livro
- Image-URL-S: URL da imagem da capa do livro (pequena)
- Image-URL-M: URL da imagem da capa do livro (média)
- Image-URL-L: URL da imagem da capa do livro (grande)

Vamos coletar apenas algumas colunas do dataset.

books = books[['ISBN',
                'Book-Title',
                  'Book-Author',
                    'Year-Of-Publication',
                      'Publisher',
                        'Image-URL-L']]

Renomeando algumas colunas

books.rename(columns = {'Book-Title':'title',
                         'Book-Author':'author',
                           'Year-Of-Publication':'year',
                             'Publisher':'publisher',
                             'Image-URL-L':'image_url'
                             }, inplace=True)

Agora temos nosso dataset contendo informações dos livros

books.head()

Dataset usuários é um conjunto de dados com 278.858 usuários e 3 atributos.

users = pd.read_csv('BX-Users.csv', sep=';',on_bad_lines='skip',encoding="latin-1", low_memory=False)
users.head()

users.shape

users.columns

- User-ID: Número de identificação do usuário
- Location: Localização do usuário
- Age: Idade do usuário

Dataset avaliações é um conjunto de dados com 1.149.780 avaliações e 3 atributos.

ratings = pd.read_csv('BX-Book-Ratings.csv', sep=';',on_bad_lines='skip',encoding="latin-1", low_memory=False)
ratings.head()

ratings.shape

1149780 registros e 3 colunas

ratings.columns

- User-ID: Número de identificação do usuário
- ISBN: Número de identificação do livro
- Book-Rating: Avaliação do livro

print(books.shape)
print(users.shape)
print(ratings.shape)

len(books["ISBN"].unique())

O fato de avaliações ter muitos registros é porque cada usuário pode ter avaliado vários livros.

ratings.rename(columns={'User-ID':'user_id','Book-Rating':'rating'},inplace=True)

Vamos ver quantas vezes os usuários avaliaram os livros

ratings['user_id'].value_counts()

Percebemos que há muitos usuários que avaliaram apenas um livro e um total de 105283 avaliações de livros diferentes.

Vamos levar em consideração apenas usuários que votaram mais de 200 vezes, pois assim os usuários que votaram apenas uma vez não terão muito peso na recomendação.

x = ratings['user_id'].value_counts()>200
x[x].shape

899 usuários fizeram mais de 200 avaliações

y = x[x].index
ratings = ratings[ratings['user_id'].isin(y)]
ratings.shape

print('Média de avaliações por usuário: ',ratings.groupby('user_id')['rating'].count().mean().round())

Agora vamos levar esses 899 usuários para o dataset de avaliações. No final, o dataset de recomendação terá seu número de registros reduzido para 526356, logicamente devido a redução de usuários. Cada usuário avaliou em média 585 livros.

Agora que temos os registros de avaliações dos usuários que avaliaram mais de 200 vezes, vamos incluir esses dados no dataframe livros, onde obteremos as informações dos livros que esses usuários avaliaram. 

ratings['ISBN'].nunique()

ratings.head()

books.shape

ratings.shape

ratings_with_books = ratings.merge(books, on='ISBN')
ratings_with_books.head()

Agora uma nova coluna foi adicionada ao dataframe, isso aumentou o seu tamanho, pois agora um livro terá mais de uma avaliação por usuário, sendo cada registro

ratings_with_books.shape

ratings_with_books["title"].nunique()

Temos agora um total de 160269 livros, pois são os livros que os usuários que avaliaram mais de 200 vezes avaliaram.

O número de registros original do dataset livros era de 271360, agora com as avaliações dos usuários que avaliaram mais de 200 vezes, o número de livros diminiu para 160269, mas o toal de registros aumentou para 487671, pois cada livro terá mais de uma avaliação por usuário.

Agora vamos obter a informação de quantas avaliações cada livro recebeu.

num_rating = ratings_with_books.groupby('title')['rating'].count().sort_values(ascending=False).reset_index()
num_rating.rename(columns={'rating':'num_of_ratings'},inplace=True)
num_rating.head()

Agora vamos incluir uma coluna de contagem de avaliações ao nosso dataframe

final_rating = ratings_with_books.merge(num_rating, on='title')
final_rating.head(2)

Vamos ver apenas livros com mais de 50 avaliações

final_rating = final_rating[final_rating['num_of_ratings']>=50]

final_rating.shape

Vamos eliminar usuários que votaram mais de uma vez no mesmo livro

final_rating.drop_duplicates(['user_id','title'],inplace=True)
final_rating.shape

2000 registros duplicados removidos

final_rating["title"].nunique(), final_rating["user_id"].nunique()

final_rating.head()

Esse é o nosso dataframe final, composto por 742 livros e 888 usuários, ou seja, 59850 registros de livros com mais de 50 avaliações, avaliados por usuários que avaliaram mais de 200 vezes.

Tabela pivot

O que faremos agora é um pivot table, onde os índices serão os usuários, as colunas serão os livros e os valores serão as avaliações. Isso porque estamos interessados em saber quais livros cada usuário avaliou, e não qual livro foi avaliado por cada usuário.

Para isso, cada coluna será baseada nos valores de user_id, os índices serão os títulos de cada livro e os valores serão as avaliações que cada usuário deu para cada livro.

book_pivot = final_rating.pivot_table(columns='user_id',index='title',values='rating')
book_pivot.fillna(0,inplace=True)
book_pivot

book_pivot.shape

Vamos agora fazer o uso da biblioteca scipy para transformar nossa matriz pivot em uma matriz esparsa. Isso porque a matriz pivot é muito grande e a maioria dos valores são nulos, então para economizar memória, vamos transformar essa matriz em uma matriz esparsa.

O método csr_matriz armazenará os dados apenas das posições que possuem valores.

from scipy.sparse import csr_matrix
books_sparse = csr_matrix(book_pivot)

Agora nossa matriz tem um tamanho reduzido e ideal para um treinamento efetivo do modelo NearestNeighbors, que se baseia na distância entre os pontos.

from sklearn.neighbors import NearestNeighbors
model = NearestNeighbors(algorithm='brute')
model.fit(books_sparse)

Vamos agora encontrar os vizinhos mais próximos de um determinado livro. Com model.kneighbors, passamos o índice do livro que queremos encontrar os vizinhos mais próximos em formato de um array 2D, onde o primeiro elemento é o índice do livro e o segundo elemento é o índice do usuário. Além disso passamos o número de vizinhos que queremos encontrar, que no caso é 6.

distance, suggestion = model.kneighbors(book_pivot.iloc[237,:].values.reshape(1,-1),n_neighbors=6)

Vamos agora criar um loop que apresenta todas as sugestões de livros para um determinado usuário.

for i in range(len(suggestion)):
    print(book_pivot.index[suggestion[i]])

Nomes dos livros presentes no dataframe

books_name = book_pivot.index

Vamos salvar o modelo, nome dos livros, o data frame final e a matriz pivot em arquivos para que possamos usar em outros notebooks.

import pickle
pickle.dump(model,open('artifacts/model.pkl','wb'))
pickle.dump(books_name,open('artifacts/books_name.pkl','wb'))
pickle.dump(final_rating,open('artifacts/final_rating.pkl','wb'))
pickle.dump(book_pivot,open('artifacts/book_pivot.pkl','wb'))

Agora vamos criar uma função geral que obtem a recomendação de livros para qualquer usuário, como nós.

def recommend_book(book_name):
    book_id = np.where(book_pivot.index==book_name)[0][0]
    distance, suggestion = model.kneighbors(book_pivot.iloc[book_id,:].values.reshape(1,-1),n_neighbors=6)
    
    recommendations = []

    for i in range(len(suggestion)):
        books = book_pivot.index[suggestion[i]]
        for j in books:
            url = final_rating.loc[final_rating['title'].str.contains(j, case=False, regex=False), 'image_url'].values[0]
            recommendation = {'Book': j, 'URL Image': url}
            recommendations.append(recommendation)

    recommendations_df = pd.DataFrame(recommendations)
    return recommendations_df

book_name = 'Harry Potter and the Chamber of Secrets (Book 2)'
recommend_book(book_name)
