==> 15-Recommender Systems/06-Recomendacao_Emprego_Dashboard.ipynb <==
!pip install dash dash_bootstrap_components

import pandas as pd
import numpy as np
import nltk
from google.colab import drive
import pickle
import dash
from dash import dcc, html, Dash
from dash.dependencies import Input, Output, State
import dash_bootstrap_components as dbc
import plotly.express as px

# Para manter montada a pasta do Google Drive no COLAB
from google.colab import drive
drive.mount('/content/drive',force_remount=True)

applicant_ids = np.array([2, 3, 6,  11,   326])

# Caminho para o arquivo pickle no Google Drive
caminho_arquivo = '/content/drive/MyDrive/Dados_Empregos/random_user_dfs1.pkl'

# Carregue os dados do arquivo pickle
with open(caminho_arquivo, 'rb') as arquivo:
    random_user_dfs1 = pickle.load(arquivo)

# Caminho para o arquivo pickle no Google Drive
caminho_arquivo2 = '/content/drive/MyDrive/Dados_Empregos/random_user_dfs2.pkl'

# Carregue os dados do arquivo pickle
with open(caminho_arquivo2, 'rb') as arquivo:
    random_user_dfs2 = pickle.load(arquivo)

def fig1(n):
  random_user_dfs1[n]['score'] = random_user_dfs1[n]['score'].astype(float)
  random_user_dfs1[n] = random_user_dfs1[n].drop_duplicates(subset='title')

  fig1 = px.bar(random_user_dfs1[n],
                  y="title",
                  x="score",
                  color="score",
                  orientation='h',  # Configuração para barras horizontais
                  color_continuous_scale=px.colors.diverging.RdYlGn,
                  text_auto=True).update_layout(bargap=0.2, bargroupgap=0.1)  # Ajuste os valores conforme necessário

  fig1.update_layout(title='Recomendação de Empregos',
                      xaxis_title='Similaridade',
                      yaxis_title='Emprego',
                      plot_bgcolor='white',
                      paper_bgcolor='white',
                      font_color='black')

  fig1.update_xaxes(gridwidth=0.5, gridcolor='#D9D9D9',range=[0, 1])  # Agora é o eixo x
  fig1.update_yaxes(showgrid=True, categoryorder="total ascending", tickfont=dict(size=10))  # Agora é o eixo y
  fig1.update_traces(showlegend=False)
  fig1.update_coloraxes(showscale=False)
  fig1
  return fig1

def fig2(n):
  random_user_dfs2[n]['score'] = random_user_dfs2[n]['score'].astype(float)
  random_user_dfs2[n] = random_user_dfs2[n].drop_duplicates(subset='title')

  fig1 = px.bar(random_user_dfs2[n],
                  y="title",
                  x="score",
                  color="score",
                  orientation='h',  # Configuração para barras horizontais
                  color_continuous_scale=px.colors.diverging.RdYlGn,
                  text_auto=True).update_layout(bargap=0.2, bargroupgap=0.1)  # Ajuste os valores conforme necessário

  fig1.update_layout(title='Recomendação de Empregos',
                      xaxis_title='Similaridade',
                      yaxis_title='Emprego',
                      plot_bgcolor='white',
                      paper_bgcolor='white',
                      font_color='black')

  fig1.update_xaxes(gridwidth=0.5, gridcolor='#D9D9D9',range=[0, 1])  # Agora é o eixo x
  fig1.update_yaxes(showgrid=True, categoryorder="total ascending", tickfont=dict(size=10))  # Agora é o eixo y
  fig1.update_traces(showlegend=False)
  fig1.update_coloraxes(showscale=False)
  fig1
  return fig1

fig1(4)

fig2(4)

app = Dash(__name__, external_stylesheets = [dbc.themes.COSMO])
server = app.server
app.layout = dbc.Container([



    html.Div(
        [
            html.H1("Dashboard - Recomendação de Emprego", style={'color': 'white'}),  # Título do cabeçalho
            html.P("Recomendação de vaga para um candidato", style={'color': 'white'}),  # Texto descritivo
        ],
        style={'background-color': 'black', 'padding': '20px'}  # Estilo do cabeçalho
    ),

    dbc.Row([

      html.H2("Análise dos Dados", style={'color': 'black', 'textAlign': 'center'}),


        dbc.Col([
            html.H3(children="Utilizando TF-IDF",style={'textAlign': 'center'}),
            dcc.Dropdown(id="botao-usuario1",
                            options=[{"label":"Escritora Voluntária Blog Uloop", "value":0},
                              {"label":"Estágio em Marketing e Preparo de Serviços e Cozinha", "value":1},
                                    {"label":"Assistente de Projeto", "value":2},
                                     {"label":"Clerk de Deli e Atendente", "value":3},
                                     {"label":"Desenvolvedor Java", "value":4}],
                         value = 1),

            dcc.Graph(id="grafico1",
                    figure = fig1(-1))




    ]),



    ], style={
            'box-shadow': '5px 5px 5px 0px rgba(0, 0, 0, 0.5)',  # Propriedades da sombra
            'padding': '20px',  # Preenchimento interno
            'margin': '20px',  # Margem externa
            'border-radius': '10px',  # Raio das bordas
        }),

    dbc.Row([
        #html.H2("Análise dos Dados", style={'color': 'black', 'textAlign': 'center'}),
        dbc.Col([html.H3(children="Utilizando Count Vectorizer",style={'textAlign': 'center'}),


            dcc.Graph(id="grafico2",

              figure = fig2(-1))])

    ], style={
            'box-shadow': '5px 5px 5px 0px rgba(0, 0, 0, 0.5)',  # Propriedades da sombra
            'padding': '20px',  # Preenchimento interno
            'margin': '20px',  # Margem externa
            'border-radius': '10px',  # Raio das bordas
        })

], fluid=True)

@app.callback(
    Output('grafico1', 'figure'),
    Output('grafico2', 'figure'),
    [Input('botao-usuario1', 'value')]
)

def mudar_graf1(valor):
  return fig1(valor),fig2(valor)



if __name__ == "__main__":
  #app.run_server()
  app.run(jupyter_mode="external")
